wip: fix coefficients for grid display

+++ src/app/make_density.hh
@@ -8,6 +8,7 @@
 #include <thread>
 #include <memory>
 #include <vector>
+#include <cmath>
 
 
 class InfoHash {
@@ -65,7 +66,7 @@ public:
     return std::thread([&](){make_grids(generation_position, running.get(),
                                         vertices, cv_generation, m);});
   }
-  static const double vector[7][3];
+  static const int vector[7][3];
 
 private:
   size_t nb_voxels;
